refactored papers
