+++ frontend/src/features/papers/types/index.ts
